Extract post-process generators from mapgen.cpp
- Moved riot damage and aftershock ruin to mapgen_post_process.cpp/.h
- Added test OMT and characterization tests for post-process behavior
