+++ src/mapgen.cpp
@@ -1,4 +1,5 @@
 #include "mapgen.h"
+#include "mapgen_post_process.h"
 
 #include <algorithm>
 #include <array>
@@ -6,7 +7,6 @@
 #include <cstdint>
 #include <functional>
 #include <limits>
-#include <list>
 #include <map>
 #include <memory>
 #include <optional>
@@ -37,7 +37,6 @@
 #include "cuboid_rectangle.h"
 #include "debug.h"
 #include "dialogue.h"
-#include "drawing_primitives.h"
 #include "enum_conversions.h"
 #include "enums.h"
 #include "field_type.h"
@@ -103,9 +102,6 @@
 #include "weighted_list.h"
 #include "magic_teleporter_list.h"
 
-static const field_type_str_id field_fd_blood( "fd_blood" );
-static const field_type_str_id field_fd_fire( "fd_fire" );
-
 static const furn_str_id furn_f_ash( "f_ash" );
 static const furn_str_id furn_f_console( "f_console" );
 static const furn_str_id furn_f_null( "f_null" );
@@ -118,8 +114,6 @@ static const furn_str_id furn_f_vending_c_off( "f_vending_c_off" );
 static const furn_str_id furn_f_vending_reinforced( "f_vending_reinforced" );
 static const furn_str_id furn_f_vending_reinforced_networked( "f_vending_reinforced_networked" );
 static const furn_str_id furn_f_vending_reinforced_off( "f_vending_reinforced_off" );
-static const furn_str_id furn_f_wreckage( "f_wreckage" );
-
 static const itype_id itype_ash( "ash" );
 static const itype_id itype_avgas( "avgas" );
 static const itype_id itype_diesel( "diesel" );
@@ -132,7 +126,6 @@ static const itype_id itype_water( "water" );
 static const mongroup_id GROUP_BREATHER( "GROUP_BREATHER" );
 static const mongroup_id GROUP_BREATHER_HUB( "GROUP_BREATHER_HUB" );
 
-static const oter_str_id oter_afs_ruins_dynamic( "afs_ruins_dynamic" );
 static const oter_str_id oter_ants_es( "ants_es" );
 static const oter_str_id oter_ants_esw( "ants_esw" );
 static const oter_str_id oter_ants_ew( "ants_ew" );
@@ -144,18 +137,10 @@ static const oter_str_id oter_ants_ns( "ants_ns" );
 static const oter_str_id oter_ants_nsw( "ants_nsw" );
 static const oter_str_id oter_ants_sw( "ants_sw" );
 static const oter_str_id oter_ants_wn( "ants_wn" );
-static const oter_str_id oter_open_air( "open_air" );
-
 static const ter_str_id ter_t_dirt( "t_dirt" );
-static const ter_str_id ter_t_floor_burnt( "t_floor_burnt" );
 static const ter_str_id ter_t_grass( "t_grass" );
-static const ter_str_id ter_t_metal_floor( "t_metal_floor" );
 static const ter_str_id ter_t_null( "t_null" );
 static const ter_str_id ter_t_pseudo_phase( "t_pseudo_phase" );
-static const ter_str_id ter_t_snow( "t_snow" );
-static const ter_str_id ter_t_snow_metal_floor( "t_snow_metal_floor" );
-static const ter_str_id ter_t_wall_burnt( "t_wall_burnt" );
-static const ter_str_id ter_t_wall_prefab_metal( "t_wall_prefab_metal" );
 
 static const trait_id trait_NPC_STATIC_NPC( "NPC_STATIC_NPC" );
 
@@ -173,540 +158,7 @@ static const vproto_id vehicle_prototype_shopping_cart( "shopping_cart" );
 
 static constexpr int MON_RADIUS = 3;
 
-enum class blood_trail_direction : int {
-    first = 1,
-    NORTH = 1,
-    SOUTH = 2,
-    EAST = 3,
-    WEST = 4,
-    last = 4
-};
-
-static tripoint_bub_ms get_point_from_direction( int direction,
-        const tripoint_bub_ms &current_tile )
-{
-    switch( static_cast<blood_trail_direction>( direction ) ) {
-        case blood_trail_direction::NORTH:
-            return tripoint_bub_ms( current_tile.x(), current_tile.y() - 1, current_tile.z() );
-        case blood_trail_direction::SOUTH:
-            return tripoint_bub_ms( current_tile.x(), current_tile.y() + 1, current_tile.z() );
-        case blood_trail_direction::WEST:
-            return tripoint_bub_ms( current_tile.x() - 1, current_tile.y(), current_tile.z() );
-        case blood_trail_direction::EAST:
-            return tripoint_bub_ms( current_tile.x() + 1, current_tile.y(), current_tile.z() );
-    }
-    // This shouldn't happen unless the function is used incorrectly
-    debugmsg( "Attempted to get point from invalid direction.  %d", direction );
-    return current_tile;
-}
-
-static bool tile_can_have_blood( map &md, const tripoint_bub_ms &current_tile,
-                                 bool wall_streak, int days_since_cataclysm )
-{
-    // Wall streaks stick to walls, like blood was splattered against the surface
-    // Floor streaks avoid obstacles to look more like a person left the streak behind (Not walking through walls, closed doors, windows, or furniture)
-    if( wall_streak ) {
-        return md.has_flag_ter( ter_furn_flag::TFLAG_WALL, current_tile ) &&
-               !md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile );
-    } else {
-        if( !md.has_flag_ter( ter_furn_flag::TFLAG_INDOORS, current_tile ) &&
-            x_in_y( days_since_cataclysm, 30 ) ) {
-            // Placement of blood outdoors scales down over the course of 30 days until no further blood is placed.
-            return false;
-        }
-
-        return !md.has_flag_ter( ter_furn_flag::TFLAG_WALL, current_tile ) &&
-               !md.has_flag_ter( ter_furn_flag::TFLAG_WINDOW, current_tile ) &&
-               !md.has_flag_ter( ter_furn_flag::TFLAG_DOOR, current_tile ) &&
-               !md.has_furn( current_tile );
-    }
-
-}
-
-static void place_blood_on_adjacent( map &md, const tripoint_bub_ms &current_tile, int chance,
-                                     int days_since_catacylsm )
-{
-    for( int i = static_cast<int>( blood_trail_direction::first );
-         i <= static_cast<int>( blood_trail_direction::last ); i++ ) {
-        tripoint_bub_ms adjacent_tile = get_point_from_direction( i, current_tile );
-
-        if( !tile_can_have_blood( md, adjacent_tile, false, days_since_catacylsm ) ) {
-            continue;
-        }
-        if( rng( 1, 100 ) < chance ) {
-            md.add_field( adjacent_tile, field_fd_blood );
-        }
-    }
-}
-
-static void place_blood_streaks( map &md, const tripoint_bub_ms &current_tile,
-                                 int days_since_cataclysm )
-{
-    int streak_length = rng( 3, 12 );
-    int streak_direction = rng( static_cast<int>( blood_trail_direction::first ),
-                                static_cast<int>( blood_trail_direction::last ) );
-
-    bool wall_streak = md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_WALL, current_tile );
-
-    if( !tile_can_have_blood( md, current_tile, wall_streak, days_since_cataclysm ) ) {
-        // Quick check the tile is valid.
-        return;
-    }
-
-    md.add_field( current_tile, field_fd_blood );
-    tripoint_bub_ms last_tile = current_tile;
-
-    for( int i = 0; i < streak_length; i++ ) {
-        tripoint_bub_ms destination_tile = get_point_from_direction( streak_direction, last_tile );
-
-        if( !tile_can_have_blood( md, destination_tile, wall_streak, days_since_cataclysm ) ) {
-            // We hit a non-valid tile. Try to find a new direction otherwise just terminate the streak.
-            bool terminate_streak = true;
-
-            for( int ii = static_cast<int>( blood_trail_direction::first );
-                 ii <= static_cast<int>( blood_trail_direction::last );
-                 ii++ ) {
-                if( ii == streak_direction ) {
-                    // We don't want to check the direction we came from. No turning around!
-                    continue;
-                }
-                tripoint_bub_ms adjacent_tile = get_point_from_direction( ii, last_tile );
-
-                if( tile_can_have_blood( md, adjacent_tile, wall_streak, days_since_cataclysm ) ) {
-                    streak_direction = ii;
-                    destination_tile = adjacent_tile;
-                    terminate_streak = false;
-                    break;
-                }
-            }
-
-            if( terminate_streak ) {
-                break;
-            }
-        }
-
-        if( rng( 1, 100 ) < 5 + i * 3 ) {
-            // Sometimes a streak should skip a tile, the chance increases with each step
-            last_tile = destination_tile;
-            continue;
-        }
-
-        if( rng( 1, 100 ) < 10 + i * 5 ) {
-            // Sometimes a streak should end early, the chance increases with each step.
-            // This is just a hack to further weight the distribution in favor of short streaks over long ones.
-            break;
-        }
-
-        md.add_field( destination_tile, field_fd_blood );
-        last_tile = destination_tile;
-
-
-        if( ( rng( 1, 100 ) < 30 + i * 3 ) && !wall_streak ) {
-            // Floor streaks can meander and the probability of meandering increases with each step
-            // Long straight streaks aren't visually interesting. So sometimes a streak will curve by meandering to the side.
-            int new_direction = 0;
-            if( streak_direction == static_cast<int>( blood_trail_direction::NORTH ) ||
-                streak_direction == static_cast<int>( blood_trail_direction::SOUTH ) ) {
-                new_direction = one_in( 2 ) ? static_cast<int>( blood_trail_direction::EAST ) : static_cast<int>(
-                                    blood_trail_direction::WEST );
-            } else {
-                new_direction = one_in( 2 ) ? static_cast<int>( blood_trail_direction::NORTH ) : static_cast<int>(
-                                    blood_trail_direction::SOUTH );
-            }
-
-            tripoint_bub_ms adjacent_tile = get_point_from_direction( new_direction, last_tile );
-            if( tile_can_have_blood( md, adjacent_tile, wall_streak, days_since_cataclysm ) ) {
-                md.add_field( adjacent_tile, field_fd_blood );
-                last_tile = adjacent_tile;
-            }
-        }
-    }
-}
-
-static void place_bool_pools( map &md, const tripoint_bub_ms &current_tile,
-                              int days_since_cataclysm )
-{
-    if( !tile_can_have_blood( md, current_tile, false, days_since_cataclysm ) ) {
-        // Quick check the first tile is valid for placement
-        return;
-    }
-
-    md.add_field( current_tile, field_fd_blood );
-    place_blood_on_adjacent( md, current_tile, 60, days_since_cataclysm );
-
-    for( int i = static_cast<int>( blood_trail_direction::first );
-         i <= static_cast<int>( blood_trail_direction::last ); i++ ) {
-        tripoint_bub_ms adjacent_tile = get_point_from_direction( i, current_tile );
-        if( !tile_can_have_blood( md, adjacent_tile, false, days_since_cataclysm ) ) {
-            continue;
-        }
-        place_blood_on_adjacent( md, adjacent_tile, 30, days_since_cataclysm );
-    }
-}
-
-struct generator_vars {
-    int scaling_days_start;
-    int scaling_days_end;
-    int num_attempts;
-    int percent_chance;
-    int min_intensity;
-    int max_intensity;
-};
-
-static void GENERATOR_bash_damage( map &md,
-                                   std::list<tripoint_bub_ms> &all_points_in_map,
-                                   int days_since_cataclysm )
-{
-    // Later, this will be loaded from json.
-    generator_vars bash_vars{};
-    bash_vars.scaling_days_start = 0; // irrelevant for this one
-    bash_vars.scaling_days_end = days_since_cataclysm; // irrelevant for this one
-    bash_vars.num_attempts = 250; // Roughly half as many attempts as old version, may need tweaking
-    bash_vars.percent_chance = 10;
-    bash_vars.min_intensity = 6; // For this generator: Bash damage
-    bash_vars.max_intensity = 60; // For this generator: Bash damage
-
-    for( int i = 0; i < bash_vars.num_attempts; i++ ) {
-        if( !x_in_y( bash_vars.percent_chance, 100 ) ) {
-            continue; // failed roll
-        }
-        const tripoint_bub_ms current_tile = random_entry( all_points_in_map );
-        if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ) {
-            continue;
-        }
-        md.bash( current_tile, rng( bash_vars.min_intensity, bash_vars.max_intensity ) );
-    }
-}
-
-static void GENERATOR_move_items( map &md,
-                                  std::list<tripoint_bub_ms> &all_points_in_map,
-                                  int days_since_cataclysm )
-{
-    // Later, this will be loaded from json.
-    generator_vars mover_vars{};
-    mover_vars.scaling_days_start = 0; // irrelevant for this one, currently.
-    mover_vars.scaling_days_end = days_since_cataclysm; // irrelevant for this one
-
-    // NOTE: Each tile of items is fully iterated over, to eliminate the effects of stack ordering.
-    // Otherwise we would be biased towards the front of the stack
-    mover_vars.num_attempts = 250; // Roughly half as many attempts as old version, may need tweaking
-
-    mover_vars.percent_chance = 10;
-    mover_vars.min_intensity = 0; // For this generator: Min distance moved. Note: NOT IMPLEMENTED
-    mover_vars.max_intensity = 3; // For this generator: Max distance moved
-
-    for( int i = 0; i < mover_vars.num_attempts; i++ ) {
-        const tripoint_bub_ms current_tile = random_entry( all_points_in_map );
-        if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ) {
-            continue;
-        }
-        auto item_iterator = md.i_at( current_tile.xy() ).begin();
-        while( item_iterator != md.i_at( current_tile.xy() ).end() ) {
-            // Some items must not be moved out of SEALED CONTAINER
-            if( md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_SEALED, current_tile ) &&
-                md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_CONTAINER, current_tile ) ) {
-                // Seed should stay in their planter for map::grow_plant to grow them
-                if( md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_PLANT, current_tile ) &&
-                    item_iterator->is_seed() ) {
-                    item_iterator++;
-                    continue;
-                }
-            }
-
-            if( x_in_y( mover_vars.percent_chance, 100 ) ) {
-                // pick a new spot...
-                tripoint_bub_ms destination_tile(
-                    current_tile.x() + rng( -mover_vars.max_intensity, mover_vars.max_intensity ),
-                    current_tile.y() + rng( -mover_vars.max_intensity, mover_vars.max_intensity ),
-                    current_tile.z() );
-                // oops, don't place out of bounds. just skip moving
-                const bool outbounds_X = destination_tile.x() < 0 || destination_tile.x() >= SEEX * 2;
-                const bool outbounds_Y = destination_tile.y() < 0 || destination_tile.y() >= SEEY * 2;
-                const bool cannot_place = md.has_flag( ter_furn_flag::TFLAG_DESTROY_ITEM, destination_tile ) ||
-                                          md.has_flag( ter_furn_flag::TFLAG_NOITEM, destination_tile ) || !md.has_floor( destination_tile );
-                if( outbounds_X || outbounds_Y || cannot_place ) {
-                    item_iterator++;
-                    continue;
-                } else {
-                    item copy( *item_iterator );
-                    // add a copy of our item to the destination...
-                    md.add_item( destination_tile, copy );
-                    // and erase the one at our source.
-                    item_iterator = md.i_at( current_tile.xy() ).erase( item_iterator );
-                }
-            } else {
-                item_iterator++;
-            }
-
-        }
-    }
-
-}
-
-static void GENERATOR_add_fire( map &md,
-                                std::list<tripoint_bub_ms> &all_points_in_map,
-                                int days_since_cataclysm )
-{
-    // Later, this will be loaded from json.
-    generator_vars fire_vars{};
-    fire_vars.scaling_days_start = 0;
-    fire_vars.scaling_days_end = 14;
-
-    // Placeholder. Number selected so that the # of fires is close to the old implementation.
-    fire_vars.num_attempts = 2;
-
-    // FIXME? I'm concerned by the fact that the initial chance scales higher the later the last scaling day is.
-    // This is not great, but it ramps down linearly without relying on magic numbers.
-    fire_vars.percent_chance = std::max( fire_vars.scaling_days_end - days_since_cataclysm, 0 );
-
-    fire_vars.min_intensity = 1; // For this generator: field intensity
-    fire_vars.max_intensity = 3; // For this generator: field intensity
-
-    for( int i = 0; i < fire_vars.num_attempts; i++ ) {
-        if( !x_in_y( fire_vars.percent_chance, 100 ) ) {
-            continue; // failed roll
-        }
-        const tripoint_bub_ms current_tile = random_entry( all_points_in_map );
-        if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ) {
-            continue;
-        }
-
-        if( x_in_y( fire_vars.percent_chance, 100 ) ) {
-            // FIXME: Magic number 3. Replace with some value loaded into generator_vars?
-            if( md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_FLAMMABLE, current_tile ) ||
-                md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_FLAMMABLE_ASH, current_tile ) ||
-                md.has_flag_ter_or_furn( ter_furn_flag::TFLAG_FLAMMABLE_HARD, current_tile ) ||
-                days_since_cataclysm < 3 ) {
-                // Only place fire on flammable surfaces unless the cataclysm started very recently
-                // Note that most floors are FLAMMABLE_HARD, this is fine. This check is primarily geared
-                // at preventing fire in the middle of roads or parking lots.
-                md.add_field( current_tile, field_fd_fire,
-                              rng( fire_vars.min_intensity, fire_vars.max_intensity ) );
-            }
-        }
-
-    }
-
-}
-
-static void GENERATOR_pre_burn( map &md,
-                                std::list<tripoint_bub_ms> &all_points_in_map,
-                                int days_since_cataclysm )
-{
-    // Later, this will be loaded from json.
-    generator_vars burnt_vars{};
-    // Fires are still raging around this time, but some start appearing
-    // Never appears before this date
-    burnt_vars.scaling_days_start = 3;
-
-    burnt_vars.scaling_days_end = 14; // Continues appearing at maximum appearance rate after this day
-    burnt_vars.num_attempts = 1; // Currently only applied to the whole map, so one pass.
-
-    burnt_vars.min_intensity = 6; // For this generator: % chance at start day
-    burnt_vars.max_intensity = 28; // For this generator: % chance at end day
-
-    // between start and end day we linearly interpolate.
-    double lerp_scalar = static_cast<double>(
-                             static_cast<double>( days_since_cataclysm - burnt_vars.scaling_days_start ) /
-                             static_cast<double>( burnt_vars.scaling_days_end - burnt_vars.scaling_days_start ) );
-    burnt_vars.percent_chance = lerp( burnt_vars.min_intensity, burnt_vars.max_intensity, lerp_scalar );
-    // static values outside that range. Note we do not use std::clamp because the chance is *0* until the start day is reached
-    if( days_since_cataclysm < burnt_vars.scaling_days_start ) {
-        burnt_vars.percent_chance = 0;
-    } else if( days_since_cataclysm >= burnt_vars.scaling_days_end ) {
-        burnt_vars.percent_chance = burnt_vars.max_intensity;
-    }
-
-    for( int i = 0; i < burnt_vars.num_attempts; i++ ) {
-        if( !x_in_y( burnt_vars.percent_chance, 100 ) ) {
-            continue; // failed roll
-        }
-        for( tripoint_bub_ms current_tile : all_points_in_map ) {
-            if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ||
-                md.has_flag_ter( ter_furn_flag::TFLAG_GOES_DOWN, current_tile ) ||
-                md.has_flag_ter( ter_furn_flag::TFLAG_GOES_UP, current_tile ) ) {
-                // skip natural underground walls, or any stairs. (Even man-made or wooden stairs)
-                continue;
-            }
-            if( md.has_flag_ter( ter_furn_flag::TFLAG_WALL, current_tile ) ) {
-                // burnt wall
-                md.ter_set( current_tile.xy(), ter_t_wall_burnt );
-            } else if( md.has_flag_ter( ter_furn_flag::TFLAG_INDOORS, current_tile ) ||
-                       md.has_flag_ter( ter_furn_flag::TFLAG_DOOR, current_tile ) ) {
-                // if we're indoors but we're not a wall, then we must be a floor.
-                // doorways also get burned to the ground.
-                md.ter_set( current_tile.xy(), ter_t_floor_burnt );
-            } else if( !md.has_flag_ter( ter_furn_flag::TFLAG_INDOORS, current_tile ) ) {
-                // if we're outside on ground level, burn it to dirt.
-                if( current_tile.z() == 0 ) {
-                    md.ter_set( current_tile.xy(), ter_t_dirt );
-                }
-            }
-
-            // destroy any furniture that is in the tile. it's been burned, after all.
-            md.furn_set( current_tile.xy(), furn_str_id::NULL_ID() );
-
-            // destroy all items in the tile.
-            md.i_clear( current_tile.xy() );
-        }
-    }
-}
-
-static void GENERATOR_riot_damage( map &md, const tripoint_abs_omt &p, bool is_a_road )
-{
-    std::list<tripoint_bub_ms> all_points_in_map;
-
-
-    int days_since_cataclysm = to_days<int>( calendar::turn - calendar::start_of_cataclysm );
-
-    // Placeholder / FIXME
-    // This assumes that we're only dealing with regular 24x24 OMTs. That is likely not the case.
-    for( int i = 0; i < SEEX * 2; i++ ) {
-        for( int n = 0; n < SEEY * 2; n++ ) {
-            tripoint_bub_ms current_tile( i, n, p.z() );
-            all_points_in_map.push_back( current_tile );
-        }
-    }
-
-    // Run sub generators associated with this generator. Currently hardcoded.
-    GENERATOR_bash_damage( md, all_points_in_map, days_since_cataclysm );
-    GENERATOR_move_items( md, all_points_in_map, days_since_cataclysm );
-    GENERATOR_add_fire( md, all_points_in_map, days_since_cataclysm );
-    // HACK: Don't burn roads to the ground! This should be resolved when the system is moved to json
-    if( !is_a_road ) {
-        GENERATOR_pre_burn( md, all_points_in_map, days_since_cataclysm );
-    }
-
-    // NOTE: Below currently only runs for bloodstains.
-    for( size_t i = 0; i < all_points_in_map.size(); i++ ) {
-        // Pick a tile at random!
-        tripoint_bub_ms current_tile = random_entry( all_points_in_map );
-
-        // Do nothing at random!;
-        if( x_in_y( 10, 100 ) ) {
-            continue;
-        }
-        // Skip naturally occuring underground wall tiles
-        if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ) {
-            continue;
-        }
-        // Set some fields at random!
-        if( x_in_y( 15, 1000 ) ) {
-            int behavior_roll = rng( 1, 100 );
-
-            if( behavior_roll <= 20 ) {
-                if( tile_can_have_blood( md, current_tile, md.has_flag_ter( ter_furn_flag::TFLAG_WALL,
-                                         current_tile ), days_since_cataclysm ) ) {
-                    md.add_field( current_tile, field_fd_blood );
-                }
-            } else if( behavior_roll <= 60 ) {
-                place_blood_streaks( md, current_tile, days_since_cataclysm );
-            } else {
-                place_bool_pools( md, current_tile, days_since_cataclysm );
-            }
-        }
-    }
-}
-
-static void GENERATOR_aftershock_ruin( map &md, const tripoint_abs_omt &p )
-{
-    std::list<tripoint_bub_ms> all_points_in_map;
-
-    // Placeholder / FIXME
-    // This assumes that we're only dealing with regular 24x24 OMTs. That is likely not the case.
-    for( int i = 0; i < SEEX * 2; i++ ) {
-        for( int n = 0; n < SEEY * 2; n++ ) {
-            tripoint_bub_ms current_tile( i, n, p.z() );
-            all_points_in_map.push_back( current_tile );
-        }
-    }
-    bool above_open_air = overmap_buffer.ter( tripoint_abs_omt( p.x(), p.y(),
-                          p.z() + 1 ) ) == oter_open_air;
-    bool above_ruined = overmap_buffer.ter( tripoint_abs_omt( p.x(), p.y(),
-                                            p.z() + 1 ) ) == oter_afs_ruins_dynamic;
-
-
-    //fully ruined
-    if( ( x_in_y( 1, 2 ) && above_open_air ) || above_ruined ) {
-        overmap_buffer.ter_set( p, oter_afs_ruins_dynamic );
-
-        for( const tripoint_bub_ms &current_tile : all_points_in_map ) {
-            if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ||
-                md.has_flag_ter( ter_furn_flag::TFLAG_GOES_DOWN, current_tile ) ||
-                md.has_flag_ter( ter_furn_flag::TFLAG_GOES_UP, current_tile ) ) {
-                // skip natural underground walls, or any stairs.
-                continue;
-            }
-            if( !one_in( 8 ) ) {
-                md.furn_set( current_tile.xy(), furn_str_id::NULL_ID() );
-            }
-            if( md.has_flag_ter( ter_furn_flag::TFLAG_WALL, current_tile ) ) {
-                !one_in( 5 ) ? md.ter_set( current_tile.xy(),
-                                           ter_t_wall_prefab_metal ) : md.ter_set( current_tile.xy(), ter_t_metal_floor );
-            }
-            if( md.has_flag_ter( ter_furn_flag::TFLAG_INDOORS, current_tile ) ) {
-                if( !one_in( 4 ) ) {
-                    p.z() == 0 ? md.ter_set( current_tile.xy(),
-                                             ter_t_snow_metal_floor ) : md.ter_set( current_tile.xy(), ter_t_snow );
-
-                }
-            }
-            if( x_in_y( 5, 1000 ) ) {
-                md.bash( current_tile, 9999, true, true );
-                draw_rough_circle( [&md, current_tile]( const point_bub_ms & p ) {
-                    if( !md.inbounds( p ) || !md.has_flag_ter( ter_furn_flag::TFLAG_FLAT, p ) ) {
-                        return;
-                    }
-                    md.bash( tripoint_bub_ms( p.x(), p.y(), current_tile.z() ), 9999, true, true );
-                }, current_tile.xy(), 3 );
-            }
-            if( !md.has_furn( current_tile.xy() ) ) {
-                md.i_clear( current_tile.xy() );
-            }
-        }
-        if( p.z() == 0 ) {
-            for( const tripoint_bub_ms &current_tile : all_points_in_map ) {
-                if( md.is_open_air( current_tile ) ) {
-                    md.ter_set( current_tile, ter_t_snow, false );
-                }
-            }
-        }
-
-        return;
-    }
-
-    //just smashed up
-    for( size_t i = 0; i < all_points_in_map.size(); i++ ) {
-
-        const tripoint_bub_ms current_tile = random_entry( all_points_in_map );
-
-        if( md.has_flag_ter( ter_furn_flag::TFLAG_NATURAL_UNDERGROUND, current_tile ) ||
-            md.is_open_air( current_tile ) ) {
-            continue;
-        }
-
-        if( x_in_y( 5, 1000 ) ) {
-            md.bash( current_tile, 9999, true, true );
-            draw_rough_circle( [&md, current_tile]( const point_bub_ms & p ) {
-                if( !md.inbounds( p ) || !md.has_flag_ter( ter_furn_flag::TFLAG_FLAT, p ) ) {
-                    return;
-                }
-                md.bash( tripoint_bub_ms( p.x(), p.y(), current_tile.z() ), 9999, true, true );
-            }, current_tile.xy(), 3 );
-        }
-
-        if( x_in_y( 5, 1000 ) ) {
-            md.bash( current_tile, 9999, true, true );
-            draw_rough_circle( [&md]( const point_bub_ms & p ) {
-                if( !md.inbounds( p ) || !md.has_flag_ter( ter_furn_flag::TFLAG_FLAT, p ) ) {
-                    return;
-                }
-                md.furn_set( p, furn_str_id::NULL_ID() );
-                md.furn_set( p, furn_f_wreckage );
-            }, current_tile.xy(), 3 );
-        }
-    }
-}
+// Post-process generators (riot damage, aftershock ruin) moved to mapgen_post_process.cpp
 
 // Assumptions:
 // - The map supplied is empty, i.e. no grid entries are in use
